# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/rooms/{id}": {
      "get": {
        "tags": [
          "room-controller"
        ],
        "operationId": "getRoomById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "room-controller"
        ],
        "operationId": "updateRoom",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Room"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "room-controller"
        ],
        "operationId": "deleteRoom",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/room-types/{roomTypeId}": {
      "put": {
        "tags": [
          "room-type-controller"
        ],
        "operationId": "updateRoomType",
        "parameters": [
          {
            "name": "roomTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomTypeRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoomType"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "room-type-controller"
        ],
        "operationId": "deleteRoomById",
        "parameters": [
          {
            "name": "roomTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rooms/hotel/{hotelId}": {
      "get": {
        "tags": [
          "room-controller"
        ],
        "operationId": "getRoomsByHotel",
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "room-controller"
        ],
        "operationId": "addRoom",
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/room-types": {
      "get": {
        "tags": [
          "room-type-controller"
        ],
        "operationId": "getAllRoomTypes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "room-type-controller"
        ],
        "operationId": "addRoomType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomTypeRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoomType"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/room-types/assets/upload": {
      "post": {
        "tags": [
          "room-type-controller"
        ],
        "operationId": "uploadAssets",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "files"
                ],
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reviews/assets/upload": {
      "post": {
        "tags": [
          "review-controller"
        ],
        "operationId": "uploadAssets_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "files"
                ],
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hotels/assets/upload": {
      "post": {
        "tags": [
          "hotel-controller"
        ],
        "operationId": "uploadAssets_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "files"
                ],
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/rooms/available/hotel/{hotelId}": {
      "get": {
        "tags": [
          "room-controller"
        ],
        "operationId": "getAvailableRooms",
        "parameters": [
          {
            "name": "hotelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/room-types/{id}": {
      "get": {
        "tags": [
          "room-type-controller"
        ],
        "operationId": "getRoomTypeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RoomType"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Amenity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "amenityName": {
            "type": "string"
          },
          "amenityIconUrl": {
            "type": "string"
          },
          "roomType": {
            "$ref": "#/components/schemas/RoomType"
          }
        }
      },
      "Asset": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "assetUrl": {
            "type": "string"
          },
          "assetThumbUrl": {
            "type": "string"
          },
          "assetType": {
            "type": "string",
            "enum": [
              "JPG",
              "JPEG",
              "PNG",
              "png"
            ]
          }
        }
      },
      "Hotel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "address": {
            "type": "string"
          },
          "contactNumber": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          },
          "roomTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomType"
            }
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          }
        }
      },
      "Location": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "pinCode": {
            "type": "string"
          }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "comment": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          }
        }
      },
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roomNumber": {
            "type": "string"
          },
          "roomType": {
            "$ref": "#/components/schemas/RoomType"
          },
          "available": {
            "type": "boolean"
          },
          "hotel": {
            "$ref": "#/components/schemas/Hotel"
          }
        }
      },
      "RoomType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "typeName": {
            "type": "string"
          },
          "capacityAdult": {
            "type": "integer",
            "format": "int32"
          },
          "capacityChild": {
            "type": "integer",
            "format": "int32"
          },
          "pricePerNight": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string"
          },
          "roomSizeInSquareFeet": {
            "type": "integer",
            "format": "int32"
          },
          "amenities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Amenity"
            }
          },
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Room"
            }
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Asset"
            }
          }
        }
      },
      "RoomTypeRequestDto": {
        "type": "object",
        "properties": {
          "typeName": {
            "type": "string"
          },
          "capacityAdult": {
            "type": "integer",
            "format": "int32"
          },
          "capacityChild": {
            "type": "integer",
            "format": "int32"
          },
          "pricePerNight": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string"
          },
          "roomSizeInSquareFeet": {
            "type": "integer",
            "format": "int32"
          },
          "amenityIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "assetIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "RoomRequestDto": {
        "type": "object",
        "properties": {
          "roomNumber": {
            "type": "string"
          },
          "roomAvailable": {
            "type": "boolean"
          },
          "hotelId": {
            "type": "integer",
            "format": "int64"
          },
          "roomTypeId": {
            "type": "integer",
            "format": "int64"
          },
          "amenityIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "assetIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
    }
  }
}